+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added support for new game installation format
+- Added 7z binary requirement
+
 ## [1.8.1] - 02.09.2024
 
 ### Fixed